feat: export portfolio table as a new component.
