+++ src/styles/index.css
@@ -16,7 +16,7 @@
 }
 
 .portfolio-primary-panel {
-  @apply w-full h-full bg-gray-300 p-2 border-r border-gray-400;
+  @apply w-full bg-gray-300 p-2 border-r border-gray-400;
 }
 
 .rdt_TableHeadRow {
